Support Min/Max/Range validations on Double/Float fields

+++ jackson/src/main/scala/com/twitter/finatra/json/internal/caseclass/validation/validators/MinValidator.scala
@@ -66,7 +66,7 @@ class MinValidator(
 
   private def validationResult(value: Number) = {
     ValidationResult.validate(
-      minValue <= value.longValue(),
+      minValue <= value.doubleValue(),
       errorMessage(value),
       errorCode(value))
   }
